fastapi-solution/etl/postgres_to_es/requirements.txt
annotated-types==0.7.0
async-timeout==4.0.3
backoff==2.2.1
black==24.4.2
certifi==2024.6.2
cfgv==3.4.0
click==8.1.7
distlib==0.3.8
elastic-transport==8.13.1
elasticsearch==8.14.0
exceptiongroup==1.2.1
filelock==3.15.3
flake8==7.1.0
flake8-isort==6.1.1
greenlet==3.0.3
identify==2.5.36
iniconfig==2.0.0
isort==5.13.2
mccabe==0.7.0
mypy-extensions==1.0.0
nodeenv==1.9.1
packaging==24.1
pathspec==0.12.1
platformdirs==4.2.2
pluggy==1.5.0
psycopg2-binary==2.9.9
pycodestyle==2.12.0
pydantic==2.7.4
pydantic-settings==2.3.4
pydantic_core==2.18.4
pyflakes==3.2.0
pytest==8.2.2
pytest-flake8==1.1.1
python-dotenv==1.0.1
PyYAML==6.0.1
redis==5.0.6
SQLAlchemy==2.0.31
tomli==2.0.1
typing_extensions==4.12.2
urllib3==2.2.2
virtualenv==20.26.2
apscheduler==3.10.4
asyncpg==0.29.0